Add x, y plots for all points. Add x, y summary plot. Change axes from 0, 1 scale to pixel scale (1920, 1080)

+++ visualize_error.py
@@ -11,6 +11,9 @@ import sys
 from get_calibration_error import get_calibration_error
 from get_calibration_folders import get_calibration_folders
 
+X_WIDTH = 1920
+Y_HEIGHT = 1080
+
 def visualize_error(root, destination):
     """
     Draws plots for data and writes results in .json format.
@@ -55,8 +58,8 @@ def visualize_error(root, destination):
         for calibration in calibs[-8:]:
             print("Processing " + subject + ": " + calibration)
             # Gaze error will be in format:
-            # [clibration_point] [ error_x[], error_y[], error_combined[], outlier_indices[] ]
-            gaze_error = get_calibration_error(calibrations_path, calibration)
+            # [calibration_point] [ error_x[], error_y[], error_combined[], outlier_indices[] ]
+            gaze_error, fixation_error = get_calibration_error(calibrations_path, calibration)
 
             # Draw and save plots. Skip this step if plots exist
             filename = subject + "_" + calibration + ".png"
@@ -89,34 +92,92 @@ def visualize_error(root, destination):
                     tmp = {}
                     filtered_x = []
                     filtered_y = []
+                    # Values as pixels
+                    pxl_x = []
+                    pxl_y = []
                     #ax = fig.add_subplot(2, 5, i + 1)
-                    ax_x = plt.subplot2grid((2, 5), (0, i))
-                    ax_y = plt.subplot2grid((2, 5), (1, i))
+                    ax_x = plt.subplot2grid((3, 5), (0, i))
+                    ax_y = plt.subplot2grid((3, 5), (1, i))
+                    ax_xy = plt.subplot2grid((3, 5), (2, i))
                     ax_x.set_title(labels[0][i])
                     ax_y.set_title(labels[1][i])
+                    ax_xy.set_title(cp_names[i])
+
                     ax_x.grid(color='lightgray', linestyle='--')
                     ax_y.grid(color='lightgray', linestyle='--')
+                    ax_xy.grid(color='lightgray', linestyle='--')
 
                     t = range(0, len(gaze_error[i][2]))
                     color = []
                     for ii in range(len(gaze_error[i][2])):
                         if ii in gaze_error[i][3]:
-                            # Outlier,
+                            # Outlier
                             color.append('r')
                         else:
                             filtered_x.append(gaze_error[i][0][ii])
+                            pxl_x.append(gaze_error[i][0][ii]*X_WIDTH)
                             filtered_y.append(gaze_error[i][1][ii])
+                            pxl_y.append(gaze_error[i][1][ii] * Y_HEIGHT)
                             color.append('b')
                     tmp["x_error"] = filtered_x
+                    tmp["x_error_pxl"] = pxl_x
                     tmp["x_stdev"] = np.std(filtered_x)
                     tmp["x_variance"] = np.var(filtered_x)
                     tmp["y_error"] = filtered_y
+                    tmp["y_error_pxl"] = pxl_y
                     tmp["y_stdev"] = np.std(filtered_y)
                     tmp["y_variance"] = np.var(filtered_y)
+
                     # plt.scatter(range(0, len(gaze_error[i][0])), gaze_error[i][0])
                     # plt.scatter(range(0, len(gaze_error[i][1])), gaze_error[i][1])
-                    ax_x.scatter(t, gaze_error[i][0], c=color)
-                    ax_y.scatter(t, gaze_error[i][1], c=color)
+                    # Convert to pixel values
+                    pxl_x_tmp = []
+                    pxl_y_tmp = []
+
+                    for val in gaze_error[i][0]:
+                        pxl_x_tmp.append(val*X_WIDTH)
+                    for val in gaze_error[i][1]:
+                        pxl_y_tmp.append(val * Y_HEIGHT)
+
+                    margin = 10
+                    # Set axis limits
+                    if abs(min(tmp['x_error_pxl'])) > abs(max(tmp['x_error_pxl'])):
+                        x_lim = abs(min(tmp['x_error_pxl']))+margin
+                    else:
+                        x_lim = abs(max(tmp['x_error_pxl']))+margin
+                    if abs(min(tmp['y_error_pxl'])) > abs(max(tmp['y_error_pxl'])):
+                        y_lim = abs(min(tmp['y_error_pxl']))+margin
+                    else:
+                        y_lim = abs(max(tmp['y_error_pxl']))+margin
+
+                    ax_xy.set_xlim(-x_lim, x_lim)
+                    ax_xy.set_ylim(-y_lim, y_lim)
+                    ax_xy.spines['left'].set_position('center')
+                    ax_xy.spines['right'].set_color(None)
+                    ax_xy.spines['bottom'].set_position('center')
+                    ax_xy.spines['top'].set_color(None)
+                    #ax_x.scatter(t, gaze_error[i][0], c=color)
+                    #ax_y.scatter(t, gaze_error[i][1], c=color)
+                    ax_x.scatter(t, pxl_x_tmp, c=color, marker='.')
+                    ax_y.scatter(t, pxl_y_tmp, c=color, marker='.')
+                    ax_xy.scatter(tmp['x_error_pxl'], tmp['y_error_pxl'], marker='.')
+
+                    # Draw fixations if exists
+                    if len(fixation_error[i]) > 0:
+                        for row in fixation_error[i]:
+                            # Correct fixation start and end frames
+                            # indexes:  0 = fixation start frame
+                            #           1 = fixation end frame
+                            #           2 = fixation error x
+                            #           3 = fixation error y
+                            #           4 = calibration point start frame
+                            #           5 = calibration point end frame
+                            fixation_t = [row[0] - row[4], row[1] - row[4]]
+                            fixation_x = [row[2]*X_WIDTH, row[2]*X_WIDTH]
+                            fixation_y = [row[3]*Y_HEIGHT, row[3]*Y_HEIGHT]
+
+                            ax_x.plot(fixation_t, fixation_x, '-k')
+                            ax_y.plot(fixation_t, fixation_y, '-k')
 
                     points[cp_names[i]] = tmp
 
@@ -183,9 +244,6 @@ def visualize_error(root, destination):
                 cp_y[i].append(value[1][i])
 
             # Copy results into json format for output
-            x_stdev = np.std(cp_x)
-            y_stdev = np.std(cp_y)
-
             tmp = { "x_error": cp_x,
                     "y_error": cp_y,
                     "x_stdev": np.std(cp_x),
@@ -205,18 +263,18 @@ def visualize_error(root, destination):
         colors = ['k', 'red', 'orange', 'c', 'blue']
         #lines = ['k-', 'darkred-', 'salmon-', 'royalblue-', 'darkblue-']
 
+        labels = ['Center',
+                  'Bottom left',
+                  'Top left',
+                  'Top right',
+                  'Bottom right']
 
         if not os.path.isfile(filepath):
-            labels = ['Center',
-                      'Bottom left',
-                      'Top left',
-                      'Top right',
-                      'Bottom right']
-
             fig = plt.figure(figsize=(20, 20))
             # Plot structure:
-            # Subplot 1 = x_error, 2 = y_error
-            ax = fig.add_subplot(2, 1, 1)
+            # Subplot 1 = x_error, 2 = y_error, 3 = errors in x, y
+            #ax = fig.add_subplot(2, 1, 1)
+            ax = plt.subplot2grid((2, 1), (0, 0))
             for i in range(5):
                 ax.set_title("x-error")
                 ax.grid(color='gray', linestyle='--', axis='y')
@@ -225,7 +283,8 @@ def visualize_error(root, destination):
                 ax.plot(x, m*x+b, color=colors[i], linestyle='-')
                 ax.legend(loc="upper left", bbox_to_anchor=(1,1))
 
-            ax = fig.add_subplot(2, 1, 2)
+            #ax = fig.add_subplot(2, 1, 2)
+            ax = plt.subplot2grid((2, 1), (1, 0))
             for i in range(5):
                 ax.set_title("y-error")
                 ax.grid(color='gray', linestyle='--', axis='y')
@@ -234,9 +293,39 @@ def visualize_error(root, destination):
                 ax.plot(x, m * x + b, color=colors[i], linestyle='-')
                 ax.legend(loc="upper left", bbox_to_anchor=(1, 1))
 
+            fig.savefig(filepath)
+            plt.close(fig)
+
+        # Save different version of summary
+
+        filename = subject + "_error_summary_xy.png"
+        filepath = os.path.join(subject_dir, filename)
+
+        if not os.path.isfile(filepath):
+            fig = plt.figure(figsize=(20,20))
+            margin = 10
+            ax = plt.subplot2grid((1, 1), (0, 0))
+            ax.spines['left'].set_position('zero')
+            ax.spines['right'].set_color(None)
+            ax.spines['bottom'].set_position('zero')
+            ax.spines['top'].set_color(None)
+            ax.grid(color='lightgray', linestyle='--')
+
+            for i in range(5):
+                # Gather points and convert to pixel values
+                cp_x_pxl = []
+                cp_y_pxl = []
+                for ii in range(len(cp_x[i])):
+                    cp_x_pxl.append(cp_x[i][ii]*X_WIDTH)
+                    cp_y_pxl.append(cp_y[i][ii]*Y_HEIGHT)
+
+                ax.scatter(cp_x_pxl, cp_y_pxl, c=colors[i], label=labels[i])
+                ax.legend(loc="upper left", bbox_to_anchor=(1, 1))
 
             fig.savefig(filepath)
             plt.close(fig)
+
+
 
         # Save summary in dictionary format
         #error_summary[subject] = json_data_calib
